Add error catching for read_sift
Each table in read_sift is now wrapped in a tryCatch - this is so that the use of `read_many_sift()` can be improved.

If the new `warn` argument is set to TRUE, the first table to fail to be read will produce a warning, and the function will return NULL. If it is FALSE (the default for `read_sift()`, an error will be thrown instead.  `read_many_sift()` and `read_many_sift2()` set `warn = TRUE` so a bad file will not disrupt the reading of a whole folder. They have been adjusted to handle NULLs while binding the data.